chore(dependencies): update multiple library versions
updated agp to 8.8.1, ksp to 2.1.20-beta2-1.0.30, androidx-navigation
to 2.8.0-alpha13, compose-multiplatform to 1.8.0-alpha03, kotlinx-datetime
to 0.6.2, ktor to 3.1.0, room to 2.7.0-beta01, firebase-bom to 33.9.0,
firebase-crashlytics-plugin to 3.0.3, and facebook-android to 18.0.2.

feat(ui): add preview for signup screen in Compose

added @Preview annotation for SignUpScreen to enable preview in
Compose. this includes a sample state for better visualization.